Add files via upload
Loops are a control flow statement that allows you to execute a block of code repeatedly until a certain condition is met. Java has three types of loops:

For loop
While loop
Do-while loop
For loop

The for loop is the most common type of loop in Java. It is used to execute a block of code a specified number of times. The syntax of the for loop is as follows:

for (initialization; condition; update) {
  // block of code
}
The initialization statement is executed once before the loop starts. The condition statement is evaluated before each iteration of the loop. If the condition is true, the block of code is executed. If the condition is false, the loop terminates. The update statement is executed after each iteration of the loop.

Here is an example of a for loop:

for (int i = 0; i < 10; i++) {
  System.out.println(i);
}
This loop will print the numbers from 0 to 9.

While loop

The while loop is used to execute a block of code repeatedly as long as a certain condition is met. The syntax of the while loop is as follows:

while (condition) {
  // block of code
}
The condition statement is evaluated before each iteration of the loop. If the condition is true, the block of code is executed. If the condition is false, the loop terminates.

Here is an example of a while loop:

int i = 0;
while (i < 10) {
  System.out.println(i);
  i++;
}
This loop will print the numbers from 0 to 9.

Do-while loop

The do-while loop is similar to the while loop, but the block of code is executed at least once, even if the condition is false. The syntax of the do-while loop is as follows:

do {
  // block of code
} while (condition);
The condition statement is evaluated after each iteration of the loop. If the condition is true, the block of code is executed. If the condition is false, the loop terminates.

Here is an example of a do-while loop:

int i = 0;
do {
  System.out.println(i);
  i++;
} while (i < 10);
This loop will print the numbers from 0 to 9, even though the condition is false after the first iteration.